refactor(time-range): split x-axis interactions into a new uPlot plugin
